Move cdr utils to a dedicated file
