+++ app/modules/cdr/endpoints_cdr.py
@@ -16,7 +16,6 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from app.core import models_core, schemas_core
 from app.core.config import Settings
 from app.core.groups.groups_type import GroupType
-from app.core.payment import schemas_payment
 from app.core.payment.payment_tool import PaymentTool
 from app.core.users import cruds_users
 from app.core.users.cruds_users import get_user_by_id, get_users
@@ -25,7 +24,6 @@ from app.dependencies import (
     get_payment_tool,
     get_settings,
     get_websocket_connection_manager,
-    hyperion_access_logger,
     is_user_a_member,
     is_user_a_member_of,
 )
@@ -34,7 +32,11 @@ from app.modules.cdr.types_cdr import (
     CdrLogActionType,
     CdrStatus,
     DocumentSignatureType,
-    PaymentType,
+)
+from app.modules.cdr.utils_cdr import (
+    check_request_consistency,
+    is_user_in_a_seller_group,
+    validate_payment,
 )
 from app.types.membership import AvailableAssociationMembership
 from app.types.module import Module
@@ -54,46 +56,6 @@ from app.utils.tools import (
     set_core_data,
 )
 
-
-async def validate_payment(
-    checkout_payment: schemas_payment.CheckoutPayment,
-    db: AsyncSession,
-) -> None:
-    paid_amount = checkout_payment.paid_amount
-    checkout_id = checkout_payment.checkout_id
-
-    checkout = await cruds_cdr.get_checkout_by_checkout_id(
-        db=db,
-        checkout_id=checkout_id,
-    )
-    if not checkout:
-        hyperion_error_logger.error(
-            f"CDR payment callback: user checkout {checkout_id} not found.",
-        )
-        raise ValueError(f"User checkout {checkout_id} not found.")  # noqa: TRY003
-
-    db_payment = models_cdr.Payment(
-        id=uuid4(),
-        user_id=checkout.user_id,
-        total=paid_amount,
-        payment_type=PaymentType.helloasso,
-    )
-    db_action = models_cdr.CdrAction(
-        id=uuid4(),
-        subject_id=checkout.user_id,
-        action_type=CdrLogActionType.payment_add,
-        action=str(checkout_payment.__dict__),
-        timestamp=datetime.now(UTC),
-    )
-    try:
-        cruds_cdr.create_payment(db=db, payment=db_payment)
-        cruds_cdr.create_action(db=db, action=db_action)
-        await db.commit()
-    except Exception:
-        await db.rollback()
-        raise
-
-
 module = Module(
     root="cdr",
     tag="Cdr",
@@ -102,98 +64,6 @@ module = Module(
 )
 
 hyperion_error_logger = logging.getLogger("hyperion.error")
-
-
-async def is_user_in_a_seller_group(
-    seller_id: UUID,
-    user: models_core.CoreUser,
-    db: AsyncSession,
-):
-    """
-    Check if the user is in the group related to a seller or CDR Admin.
-    """
-    seller = await cruds_cdr.get_seller_by_id(db, seller_id=seller_id)
-
-    if not seller:
-        raise HTTPException(
-            status_code=404,
-            detail="Seller not found.",
-        )
-
-    if is_user_member_of_an_allowed_group(
-        user=user,
-        allowed_groups=[str(seller.group_id), GroupType.admin_cdr],
-    ):
-        return user
-
-    hyperion_access_logger.warning(
-        "Is_user_a_member_of: Unauthorized, user is not a seller",
-    )
-
-    raise HTTPException(
-        status_code=403,
-        detail="Unauthorized, user is not in this seller group.",
-    )
-
-
-async def check_request_consistency(
-    db: AsyncSession,
-    seller_id: UUID | None = None,
-    product_id: UUID | None = None,
-    variant_id: UUID | None = None,
-    document_id: UUID | None = None,
-) -> models_cdr.CdrProduct | None:
-    """
-    Check that given ids are consistent, ie. product's seller_id is the given seller_id.
-    """
-    db_product: models_cdr.CdrProduct | None = None
-    if seller_id:
-        db_seller = await cruds_cdr.get_seller_by_id(db=db, seller_id=seller_id)
-        if not db_seller:
-            raise HTTPException(
-                status_code=404,
-                detail="Invalid seller_id",
-            )
-    if product_id:
-        db_product = await cruds_cdr.get_product_by_id(db=db, product_id=product_id)
-        if not db_product:
-            raise HTTPException(
-                status_code=404,
-                detail="Invalid product_id",
-            )
-        if seller_id and seller_id != db_product.seller_id:
-            raise HTTPException(
-                status_code=403,
-                detail="Product is not related to this seller.",
-            )
-    if variant_id:
-        db_variant = await cruds_cdr.get_product_variant_by_id(
-            db=db,
-            variant_id=variant_id,
-        )
-        if not db_variant:
-            raise HTTPException(
-                status_code=404,
-                detail="Invalid variant_id",
-            )
-        if product_id and product_id != db_variant.product_id:
-            raise HTTPException(
-                status_code=403,
-                detail="Variant is not related to this product.",
-            )
-    if document_id:
-        db_document = await cruds_cdr.get_document_by_id(db=db, document_id=document_id)
-        if not db_document:
-            raise HTTPException(
-                status_code=404,
-                detail="Invalid document_id",
-            )
-        if seller_id and seller_id != db_document.seller_id:
-            raise HTTPException(
-                status_code=403,
-                detail="Document is not related to this seller.",
-            )
-    return db_product
 
 
 @module.router.get(
